+++ appliance/config/bundles.php
@@ -19,5 +19,4 @@ return [
     Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['dev' => true, 'test' => true],
     FriendsOfBehat\SymfonyExtension\Bundle\FriendsOfBehatSymfonyExtensionBundle::class => ['test' => true],
     Zenstruck\Messenger\Test\ZenstruckMessengerTestBundle::class => ['test' => true],
-    Teknoo\Space\Extensions\Enterprise\Bundle\TeknooSpaceEnterpriseBundle::class => ['all' => true],
 ];